pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Matt Finnell", email = "mattfinnell104@gmail.com"},
]
dependencies = [
    "django",
    "djangorestframework",
    "django-rq",
    "python-dotenv",
    "psycopg",
    "easyocr",
    "zxing-cpp>=2.1.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "None"}

[tool.pdm]
[[tool.pdm.source]]
url = "https://pypi.org/simple"
verify_ssl = true
name = "pypi"

[tool.pdm.dev-dependencies]
dev = []

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
